requirements.txt

Django==4.0.6
django-cors-headers==3.13.0
djangorestframework==3.13.1
python-dotenv==1.0.0
requests==2.28.2

